+++ infrastructure/entry-points/reactive-web/src/main/java/co/com/reportes/api/dto/ReporteSolicitudesDto.java
@@ -2,6 +2,8 @@ package co.com.reportes.api.dto;
 
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @AllArgsConstructor
 @NoArgsConstructor
 @Getter
@@ -10,5 +12,5 @@ import lombok.*;
 public class ReporteSolicitudesDto {
     private String id;
     private Long contador;
-    private Long monto;
+    private BigDecimal monto;
 }
